add a new background
